pyproject.toml
[project]
name = "coderboys-payment-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.4",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "dotenv>=0.9.9",
    "gunicorn>=25.3.0",
    "psycopg2-binary>=2.9.12",
    "requests>=2.33.1",
    "uvicorn>=0.46.0",
    "whitenoise>=6.12.0",
]
